Refactor: #40 separate checkbox and filter result
